Fix memory corruption in PCB view
The first issue was that items bounding rectangle was not fully
inside of maxRect that is bounding PCB board. The reason is that
in removeOffBoardAnd() function only non-intersecting connectors
were removed from to-autoroute list, so they could partially be
outside of board bounding rectangle, which led to memory corruption.

The second issue was that even though some connectors were removed
from the connectors list by the removeOffBoardAnd(), they were
eventually added to the subnets, because lookup of on-the-same-wire
connectors were performed in collectEqualPotential() function.