WIP: Refactor project structure
New main project file is fritzing.pro, not phoenix.pro.
It uses subfolders with libraries, instead of including all files with
relative paths.

The idea is to
1. stay closer to Qt recommendations
2. have a stronger separation of concerns
3. prepare to add unit tests

It is not yet clear if this will work out. There might be
to many circular dependcies, so not all subdirs can be made
into libraries.