Fix issue #3683
This fixes the following use case:

Add two parts into an empty project,
one of them with a bendable leg (for example a resistor)
Go to the PCB view, connect the parts.
Go to the breadboard view. You see a ratsnest line
for the unrouted connection. Grab the bendable leg
at its endpoint and move it.

The connection is separated in breadboard view,
and not in PCB view, resulting in the view being
out of sync. The fix will keep this kind
of connections. The same applies for schematic view
instead of PCV view.