Update the timer for swapping parts
This change fixes issue #1431. I am not sure that this is the right way of fixing this issue, but at least it will mitigate the issue.
I would disable the change of the properties by scrolling, except that the property combobox had been selected. This will avoid the fact that you can unintentionally change properties of the part when scrolling through the inspector pane.